requirements.txt
django~=4.0.5
psycopg2-binary

django-autoslug~=1.9.8